// AdminPanel-2025-dev-origin/src/types/bootcamp.ts
export interface BootcampType {
  id: string | number;
  name: string;
  date: string;
  total_capacity: number;

  
  forms?: unknown[]; 
  attendees?: unknown[]; 

  
  [key: string]: unknown;
}

export interface BootcampResponse {
  success: boolean;
  message: string;
  status: number;
  data: BootcampType[];
}

export type BootcampRequest = Omit<BootcampType, "id">;